% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lsq.R
\name{calc_lsq}
\alias{calc_lsq}
\title{calc_lsq}
\usage{
calc_lsq(prop_votes, prop_seats)
}
\arguments{
\item{prop_votes}{A vector of party seats proportions.}
}
\value{
A vector of length 1 with the least-squares index.
}
\description{
Calculates Gallagher's index of proportionality or "least squares" index.
}
\examples{
camara \%>\%
    group_by(congress) \%>\%
    summarize(lsq = calc_lsq(prop_votes, prop_seats))
}
