% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nep.R
\name{calc_nep}
\alias{calc_nep}
\title{calc_nep}
\usage{
calc_nep(prop_votes, index = "lt")
}
\arguments{
\item{prop_votes}{A vector of party vote proportions.}

\item{index}{Formula for the index. Can be either "lt" for Laakso-Taagepera (1979)
or "g" for Golosov (2010).}
}
\value{
A vector of length 1 with the ENP index.
}
\description{
Calculates the effective number of parties (ENP, seats or votes),
according to two different formulae.
}
\examples{
data(camara)
camara \%>\%
    group_by(congress) \%>\%
    summarize(nep_lt = calc_nep(prop_votes),
              nep_g = calc_nep(prop_votes, index = "g"))
}
