% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_inp.R
\name{calc_pni}
\alias{calc_pni}
\title{calc_pni}
\usage{
calc_pni(prop_votes)
}
\arguments{
\item{prop_votes}{A vector of party vote proportions.}
}
\value{
A vector of length 1 with the PNI.
}
\description{
Calculates the party nationalization index (PNI).
}
\examples{
data(camara_dept)

# apply function to one party
camara_dept \%>\%
 filter(party == "ALIANZA VERDE") \%>\%
 summarize(inp = calc_pni(prop_votes_party))

# apply function to all parties
camara_dept \%>\%
  filter(party != "VOTO EN BLANCO") \%>\%
  group_by(party) \%>\%
  summarize(inp = calc_pni(prop_votes_party))
}
