% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ideal_params.R
\name{extract_ideal_params}
\alias{extract_ideal_params}
\title{extract_ideal_params}
\usage{
extract_ideal_params(model)
}
\arguments{
\item{model}{An `ideal` model object}
}
\value{
A tibble with means and confidence intervals for every difficulty and discrimination parameter
}
\description{
Extracts difficulty and discrimination parameters from an
\code{ideal} object
}
