% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fortou.R
\name{theme_fortou}
\alias{theme_fortou}
\title{theme_fortou}
\usage{
theme_fortou()
}
\value{

}
\description{
Minimal ggplot2 theme
}
\examples{
\dontrun{
library(ggplot)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "A title", subtitle = "A subtitle", x = "Weight", y = "MPG") +
    theme_fortou()

set_theme(theme_fortou())

}
}
